kotlin migration: compiles and runs

+++ src/main/java/akostenko/aicars/drawing/Line.kt
@@ -4,4 +4,4 @@ import akostenko.aicars.math.Decart
 
 import org.newdawn.slick.Color
 
-class Line(val from: Decart, val to: Decart, val color: Color, val width: Int)
+class Line(val from: Decart, val to: Decart, val color: Color, val width: Float)
